refactor(extractor): move block stats to primitives
